pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tp53-sgrna-prioritization"
version = "1.0.0"
description = "Cancer-type-stratified CRISPR-Cas9 sgRNA prioritization for TP53 hotspots."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Loke Srinivas" }
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[project.scripts]
tp53-sgrna = "src.cli:main"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"
